word-cloud: renamed files according to the project's naming patterns
