+++ components/home-components/sectors-section/word-cloud/index.js
