+++ components/home-components/sectors-section/word-cloud/tag.js
